+++ packages/implementations/react/src/atomic/Input/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Leading";
+export * from "./Password";
+export * from "./Search";