+++ packages/implementations/react/src/atomic/Input/src/input.definitions.ts
@@ -1,0 +1,7 @@
+export const NEUTRAL_COLORS = ["brand", "info", "neutral"];
+
+export const getColorButtonClose = {
+  brand: "neutral",
+  info: "neutral",
+  neutral: "neutral",
+} as { [key: string]: "neutral" };