+++ packages/implementations/react/src/atomic/Input/src/input.types.ts
@@ -1,0 +1,27 @@
+import { InputHTMLAttributes, ReactNode } from "react";
+import { InputVariants } from "@rarui/styles";
+import { Leading, Password, Search } from "./components";
+
+export interface InputComponents {
+  Leading: typeof Leading;
+  Password: typeof Password;
+  Search: typeof Search;
+}
+
+export interface Input {
+  /**
+   * Componet to be displayed on start input.
+   * @TJS-type React.ReactNode
+   */
+  leadingStart?: ReactNode;
+  /**
+   * Componet to be displayed on end input.
+   * @TJS-type React.ReactNode
+   */
+  leadingEnd?: ReactNode;
+}
+
+export type InputProperties = Input & InputVariants;
+
+export type _InputProps = InputProperties &
+  Omit<InputHTMLAttributes<HTMLInputElement>, "size">;