+++ packages/implementations/react/src/atomic/Toast/src/components/Provider/provider.stories.tsx
@@ -1,27 +1,23 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
+import { argTypesConvert } from ".storybook/utils";
 
 import { Button } from "@rarui-react/button";
 import { Box } from "@rarui-react/box";
+import docs from "../../toast.docs.json";
 
 import { Toast } from "../../Toast";
 import { useToast } from "../../hooks";
+
+const providerdDocs = docs.subComponents.find(
+  (subComponent) => subComponent.name === "Toast.Provider",
+);
 
 const meta: Meta<typeof Toast.Provider> = {
   title: "Atomic/Toast.Provider",
   component: Toast.Provider,
   argTypes: {
-    placement: {
-      control: "radio",
-      options: [
-        "bottomRight",
-        "bottomLeft",
-        "bottomCenter",
-        "topCenter",
-        "topRight",
-        "topLeft",
-      ],
-    },
+    ...argTypesConvert(providerdDocs),
     children: { control: { type: "text" } },
   },
   tags: ["autodocs"],