+++ packages/styles/src/packages/atomic/iconButton/rarui-iconButton.css.ts
@@ -258,16 +258,6 @@ const iconButton = recipe({
         backgroundColor: varsThemeBase.colors.surface.secondary,
       },
     },
-    {
-      variants: {
-        appearance: "brand",
-        variant: "tonal",
-      },
-      style: {
-        ...buttonVariants.text.brand,
-        backgroundColor: varsThemeBase.colors.surface["invert-secondary"],
-      },
-    },
   ],
 });