+++ packages/styles/src/packages/atomic/input/index.ts
@@ -1,0 +1,6 @@
+import { styles } from "./raui-input.css";
+
+export const input = {
+  classnames: { ...styles },
+};
+export type { InputVariants } from "./rarui-input.types";