+++ packages/styles/src/packages/atomic/label/index.ts
@@ -1,0 +1,6 @@
+import { styles } from "./raui-label.css";
+
+export const label = {
+  classnames: { ...styles },
+};
+export type { LabelVariants } from "./rarui-label.types";