+++ packages/styles/src/packages/atomic/link/rarui-link.css.ts
@@ -29,7 +29,7 @@ const link = recipe({
   },
   variants: {
     appearance: {
-      default: {
+      brand: {
         color: varsThemeBase.colors.content.brand,
       },
       neutral: {