+++ packages/implementations/react/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-05-01 `1.16.0`
+
+#### 🎉 New features
+
+- Added new atomic `Input` component. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `Label` component. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🐛 Bug fixes
+
+- Configured default placement in `Toast.provider` component. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-29 `1.15.0`
 
 #### 🎉 New features