+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 RarUI Styles deprive all styles needed to build components.
 
+## 2024-05-01 `1.15.0`
+
+#### 🎉 New features
+
+- Added new style pack for `input` component. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new style pack for `label` component. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🛠 Breaking changes
+
+- Renamed the `default` appearance name to `brand`. ([#30](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/30) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-29 `1.14.0`
 
 #### 🎉 New features